refactor: moved cdf.extraction_pipelines into root of utils module
